move tag filter stuff around, some of this code is ass lol
