+++ src/constants.ts
@@ -1,35 +1,1 @@
 export const DECAL_MAX_RESULTS = 100;
-
-export const GTSPORT_TAGS_MAP: { [key: string]: string } = {
-    "131": "Shapes",
-    "132": "Lines",
-    "133": "Patterns",
-    "134": "Logos",
-    "135": "Illustrations",
-    "136": "Characters",
-    "137": "Decals",
-    "138": "Others",
-    "141": "Can be Modified",
-    "142": "Cannot be Modified",
-    "143": "Can be Coloured",
-    "144": "Cannot be Coloured",
-    // "151": "White",
-    // "152": "Silver",
-    // "153": "Black",
-    // "154": "Blue",
-    // "155": "Red",
-    // "156": "Green",
-    // "157": "Grey",
-    // "158": "Yellow",
-    // "159": "Orange",
-    // "160": "Purple",
-    // "161": "Pink",
-    // "162": "Brown",
-    // "163": "Others",
-    // "171": "Replicas",
-    // "172": "Basic",
-    // "173": "Gorgeous",
-    // "174": "Pop",
-    // "175": "Retro",
-    // "176": "Unique"
-};
